+++ config/user_programs/demo_ctypes/python_funcs.py
@@ -1,4 +1,4 @@
-from ctypes import c_int, POINTER
+from ctypes import c_float, c_int, POINTER
 from lizard import user_prog_resources
 
 
@@ -100,7 +100,7 @@ class GlobalState(user_prog_resources.EncodableStructure):
     _fields_ = [
         ('done', c_int),
         ('iteration', c_int),
-        ('values', POINTER(c_int)),
+        ('values', POINTER(c_float)),
     ]
     aux_field_names = ('values',)
 
@@ -141,7 +141,7 @@ class GlobalState(user_prog_resources.EncodableStructure):
 
 class AggregationResult(user_prog_resources.EncodableStructure):
     _fields_ = [
-        ('values', POINTER(c_int)),
+        ('values', POINTER(c_float)),
     ]
     aux_field_names = ('values',)
 
@@ -183,7 +183,7 @@ class AggregationResult(user_prog_resources.EncodableStructure):
 class Dataset(user_prog_resources.EncodableStructure):
     _fields_ = [
         ('num_points', c_int),
-        ('points', POINTER(POINTER(c_int))),
+        ('points', POINTER(POINTER(c_float))),
     ]
     aux_field_names = ('points',)
